Dissolve `mod decoder` into crate root
