+++ qemu-run/src/main.rs
@@ -10,7 +10,7 @@ use std::{
 };
 
 use anyhow::{anyhow, bail};
-use defmt_decoder::decoder::{DecodeError, Table};
+use defmt_decoder::{DecodeError, Table};
 use process::Child;
 
 fn main() -> Result<(), anyhow::Error> {
@@ -92,7 +92,7 @@ fn notmain() -> Result<Option<i32>, anyhow::Error> {
 
 fn decode(frames: &mut Vec<u8>, table: &Table) -> Result<(), DecodeError> {
     loop {
-        match defmt_decoder::decoder::decode(&frames, &table) {
+        match defmt_decoder::decode(&frames, &table) {
             Ok((frame, consumed)) => {
                 println!("{}", frame.display(true));
                 let n = frames.len();
